Use UIDiffableDataSource.apply to perform diff and only update what's changed (with animation)
Since the diffable data source diffing happens on its own dedicated queue (not the main queue) that may also run work on the main thread, we had to replace the `ImmediateWhenOnMainQueueScheduler` with the `ImmediateWhenOnMainThreadScheduler`.

Also, we had to rearrange the setup of some tests since the diffable data source eagerly preloads cells.